login
did a couple of things to enable login functionaliy.

created the login modal that pops up when there is a currentUser in session.

while the user is in session the navbar changes to incude different routes within the profile tab.

logout functionality was also implemented using the signOut functionality of next